refactor(upgrades): align v4_3 with mainnet, move costaking reset to v4_4
The v4.3 upgrade on mainnet was run with the release/v4.3.x patched binary,
which only does (1) module migrations and (2) the phantom stake-expansion
remediation. v4.4.x's v4_3 handler used to do something different (costaking
rewards tracker reset + epoch boundary validation), which would diverge from
mainnet state during replay-from-genesis.

Align v4.4.x's v4_3 handler with what actually ran on mainnet:

* `app/upgrades/v4_3/upgrade.go`: replace with the phantom-remediation handler
  from release/v4.3.x. Run migrations, then RemediatePoisonedStakeExpansions,
  return.
* Add a previous-stake covenant-quorum gate to RemediatePoisonedStakeExpansions
  (Codex review P1): only force-unbond a parent when the child's
  PreviousStkCovenantSigs meet the parent's CovenantQuorum, resolved from the
  parent's own ParamsVersion. Without prev-stake quorum the child cannot have
  produced a valid spending witness on the parent's UTXO, so the parent was
  not actually spent on Bitcoin and remediation must not touch it.
* Move the phantom-remediation tests from `app/upgrades/v4_4/upgrade_test.go`
  to `app/upgrades/v4_3/upgrade_test.go` and retarget them at the v4_3
  package. Add `TestRemediate_SkipsChildWithoutPreviousStkCovenantQuorum` to
  cover the new gate. Update the in-test helper `markStakeExpansionOf` to
  populate `PreviousStkCovenantSigs` with quorum-meeting dummies so the
  existing tests still exercise the remediation path.
* Move the costaking-rewards-tracker reset, ActiveBaby zero/recalculate, and
  epoch boundary validation out of v4_3 into a new v4.4 upgrade.
  `app/upgrades/v4_4/upgrade.go` replaces the phantom-remediation stub I had
  there. Move the costaking-reset test from
  `app/upgrades/v4_3/upgrades_costaking_test.go` to
  `app/upgrades/v4_4/upgrades_costaking_test.go` and retarget at v4_4.
* Revert the BtcUndelegate / IterateBTCDelegations doc updates: v4_4 no
  longer calls those helpers, so the consumer list is back to v4_3 only and
  the removal TODOs gate on v4.3 retiring alone.
