+++ app/upgrades/v4_3/upgrade.go
@@ -2,351 +2,281 @@ package v4_3
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
-	"cosmossdk.io/collections"
-	corestoretypes "cosmossdk.io/core/store"
-	"cosmossdk.io/math"
 	store "cosmossdk.io/store/types"
 	upgradetypes "cosmossdk.io/x/upgrade/types"
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/runtime"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
-	"github.com/cosmos/cosmos-sdk/types/query"
-	stkkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	stktypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	"github.com/babylonlabs-io/babylon/v4/app/keepers"
 	"github.com/babylonlabs-io/babylon/v4/app/upgrades"
-	"github.com/babylonlabs-io/babylon/v4/app/upgrades/epoching"
-	costkkeeper "github.com/babylonlabs-io/babylon/v4/x/costaking/keeper"
-	costktypes "github.com/babylonlabs-io/babylon/v4/x/costaking/types"
-	epochingkeeper "github.com/babylonlabs-io/babylon/v4/x/epoching/keeper"
+	btcstkkeeper "github.com/babylonlabs-io/babylon/v4/x/btcstaking/keeper"
+	btcstktypes "github.com/babylonlabs-io/babylon/v4/x/btcstaking/types"
 )
 
-const UpgradeName = "v4.3"
+const (
+	UpgradeName = "v4.3"
+
+	// EventTypePhantomStakeExpansionRemediated is emitted once per parent
+	// delegation that is force-unbonded by the GHSA-4rm2-cj74-f62h remediation.
+	// External indexers can use it to reconcile prior phantom voting power.
+	EventTypePhantomStakeExpansionRemediated = "phantom_stake_expansion_remediated"
+
+	AttrParentStakingTxHash = "parent_staking_tx_hash"
+	AttrChildStakingTxHash  = "child_staking_tx_hash"
+)
 
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
-	StoreUpgrades: store.StoreUpgrades{
-		Added:   []string{},
-		Deleted: []string{},
-	},
+	StoreUpgrades:        store.StoreUpgrades{},
 }
 
-func CreateUpgradeHandler(mm *module.Manager, configurator module.Configurator, keepers *keepers.AppKeepers) upgradetypes.UpgradeHandler {
-	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-		// Run migrations before applying any other state changes.
-		migrations, err := mm.RunMigrations(ctx, configurator, fromVM)
+func CreateUpgradeHandler(mm *module.Manager, cfg module.Configurator, k *keepers.AppKeepers) upgradetypes.UpgradeHandler {
+	return func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		sdkCtx := sdk.UnwrapSDKContext(ctx)
+		logger := sdkCtx.Logger().With("upgrade", UpgradeName)
+
+		logger.Info("running module migrations")
+		vm, err := mm.RunMigrations(ctx, cfg, fromVM)
 		if err != nil {
+			logger.Error("module migrations failed", "error", err.Error())
 			return nil, err
 		}
 
-		// Validate that we are at the second block of the epoch.
-		// The second block is required because GetCurrentValidatorSet depends
-		// on InitValidatorSet which runs in the epoching BeginBlocker on the
-		// first block of the epoch.
-		if err := epoching.ValidateSecondBlockOfEpoch(ctx, keepers.EpochingKeeper); err != nil {
-			return nil, fmt.Errorf("epoch boundary validation failed: %w", err)
-		}
-
-		costkStoreKey := keepers.GetKey(costktypes.StoreKey)
-		if costkStoreKey == nil {
-			return nil, errors.New("invalid costaking types store key")
-		}
-		coStkStoreService := runtime.NewKVStoreService(costkStoreKey)
-
-		// Reset co-staker rewards tracker for ActiveBaby
-		if err := ResetCoStakerRwdsTrackerActiveBaby(
-			ctx,
-			keepers.EncCfg.Codec,
-			coStkStoreService,
-			keepers.EpochingKeeper,
-			keepers.StakingKeeper,
-			keepers.CostakingKeeper,
-		); err != nil {
+		logger.Info("scanning BTC staking store for poisoned stake-expansion parents (GHSA-4rm2-cj74-f62h)")
+		remediated, err := RemediatePoisonedStakeExpansions(sdkCtx, k.BTCStakingKeeper)
+		if err != nil {
+			logger.Error("stake-expansion remediation failed — upgrade aborting",
+				"error", err.Error())
 			return nil, err
 		}
-
-		return migrations, nil
+		logger.Info("upgrade complete", "remediated_parents", len(remediated))
+		return vm, nil
 	}
 }
 
-// ResetCoStakerRwdsTrackerActiveBaby resets the ActiveBaby in costaker rewards tracker
-// It recalculates ActiveBaby for all BABY stakers based on current delegations to active validators
-func ResetCoStakerRwdsTrackerActiveBaby(
-	ctx context.Context,
-	cdc codec.BinaryCodec,
-	costkStoreService corestoretypes.KVStoreService,
-	epochingKeeper epochingkeeper.Keeper,
-	stkKeeper *stkkeeper.Keeper,
-	coStkKeeper costkkeeper.Keeper,
-) error {
-	sb := collections.NewSchemaBuilder(costkStoreService)
-	rwdTrackers := collections.NewMap(
-		sb,
-		costktypes.CostakerRewardsTrackerKeyPrefix,
-		"costaker_rewards_tracker",
-		collections.BytesKey,
-		codec.CollValue[costktypes.CostakerRewardsTracker](cdc),
+// RemediatePoisonedStakeExpansions scans the BTC staking store for
+// stake-expansion children that were unbonded early before their inclusion
+// proof was added (the GHSA-4rm2-cj74-f62h attack pattern) and force-unbonds
+// the corresponding parent delegations. The child's own staking transaction
+// is the canonical proof that the parent's UTXO was spent on Bitcoin, so it
+// is recorded as the parent's `DelegatorUnbondingInfo.SpendStakeTx`.
+//
+// The remediation is intentionally idempotent: parents that are already
+// unbonded (legitimately or by a previous run) are skipped. Returns the
+// staking-tx hashes (as hex strings) of remediated parents in iteration
+// order.
+func RemediatePoisonedStakeExpansions(
+	ctx sdk.Context,
+	bk btcstkkeeper.Keeper,
+) ([]string, error) {
+	logger := ctx.Logger().With("upgrade", UpgradeName, "module", "stake_expansion_remediation")
+
+	type pair struct {
+		parent     *btcstktypes.BTCDelegation
+		spendTxBz  []byte
+		childHash  string
+		parentHash string
+	}
+	// Collect first, mutate after — avoids invalidating the store iterator.
+	var toRemediate []pair
+	// Dedupe by parent hash. On Bitcoin, only ONE spender of the parent's
+	// UTXO can ever confirm; on babylon we may have registered multiple
+	// stake expansions of the same ACTIVE parent before any of them
+	// confirmed. We only need to mark the parent UNBONDED once, so dedupe
+	// here and record the first encountered child as SpendStakeTx.
+	seenParent := make(map[string]struct{})
+
+	var (
+		scanned                    int
+		poisonedChildren           int
+		skippedParentMissing       int
+		skippedParentAlreadyUnbond int
+		skippedParentNotActive     int
+		skippedDuplicateParent     int
+		skippedNoPrevStkQuorum     int
 	)
 
-	// Zero out ActiveBaby in all existing rewards trackers and track previous values
-	accsWithActiveBaby, err := zeroOutCoStakerRwdsActiveBaby(ctx, rwdTrackers)
-	if err != nil {
-		return err
-	}
-
-	params := coStkKeeper.GetParams(ctx)
-	endedPeriod, err := coStkKeeper.IncrementRewardsPeriod(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Recalculate ActiveBaby for all BABY stakers based on current delegations to active validators
-	if err := updateBABYStakersRwdTracker(ctx, endedPeriod, rwdTrackers, accsWithActiveBaby, epochingKeeper, stkKeeper, coStkKeeper, params); err != nil {
-		return err
-	}
-
-	totalScore, err := getTotalScore(ctx, rwdTrackers)
-	if err != nil {
-		return err
-	}
-
-	currentRwd, err := coStkKeeper.GetCurrentRewards(ctx)
-	if err != nil {
-		return err
-	}
-
-	currentRwd.TotalScore = totalScore
-	if err := currentRwd.Validate(); err != nil {
-		return err
-	}
-
-	return coStkKeeper.SetCurrentRewards(ctx, *currentRwd)
-}
-
-type ActiveBabyTracked struct {
-	PreviousActiveBaby math.Int
-	CurrentActiveBaby  math.Int
-}
-
-// updateBABYStakersRwdTracker retrieves all BABY stakers delegating to active validators and updates their ActiveBaby
-func updateBABYStakersRwdTracker(
-	ctx context.Context,
-	period uint64,
-	rwdTrackers collections.Map[[]byte, costktypes.CostakerRewardsTracker],
-	accsWithActiveBaby map[string]ActiveBabyTracked,
-	epochingKeeper epochingkeeper.Keeper,
-	stkKeeper *stkkeeper.Keeper,
-	coStkKeeper costkkeeper.Keeper,
-	params costktypes.Params,
-) error {
-	// Get all BABY stakers delegating to active validators in the current epoch
-	babyStakers, err := getAllBABYStakers(ctx, epochingKeeper, stkKeeper)
-	if err != nil {
-		return fmt.Errorf("failed to get all BABY stakers: %w", err)
-	}
-
-	// Add current BABY delegations to the tracking map
-	for delegatorAddr, babyAmount := range babyStakers {
-		data, found := accsWithActiveBaby[delegatorAddr]
-		if found {
-			data.CurrentActiveBaby = data.CurrentActiveBaby.Add(babyAmount)
-			accsWithActiveBaby[delegatorAddr] = data
-		} else {
-			accsWithActiveBaby[delegatorAddr] = ActiveBabyTracked{
-				PreviousActiveBaby: math.ZeroInt(),
-				CurrentActiveBaby:  babyAmount,
-			}
+	if err := bk.IterateBTCDelegations(ctx, func(child *btcstktypes.BTCDelegation) error {
+		scanned++
+		if !child.IsStakeExpansion() {
+			return nil
 		}
-	}
-
-	// Update the costaker rewards trackers for all accounts
-	for accAddrStr, babyTracked := range accsWithActiveBaby {
-		if err := updateCostakerActiveBabyRewardsTracker(
-			ctx,
-			coStkKeeper,
-			period,
-			rwdTrackers,
-			sdk.MustAccAddressFromBech32(accAddrStr),
-			params,
-			babyTracked,
-		); err != nil {
-			return err
+		if child.HasInclusionProof() {
+			return nil
 		}
-	}
-
-	return nil
-}
-
-// getAllBABYStakers retrieves all BABY stakers by iterating over the current epoch's active validators
-func getAllBABYStakers(ctx context.Context, epochingKeeper epochingkeeper.Keeper, stkKeeper *stkkeeper.Keeper) (map[string]math.Int, error) {
-	stkQuerier := stkkeeper.NewQuerier(stkKeeper)
-	babyStakers := make(map[string]math.Int)
-
-	// Get the current epoch's validator set from epoching keeper
-	valSet := epochingKeeper.GetCurrentValidatorSet(ctx)
-
-	// Iterate over validators in the current epoch
-	for _, val := range valSet {
-		valAddr := sdk.ValAddress(val.Addr)
-
-		// Get all delegations for this active validator
-		if err := getValidatorDelegations(ctx, stkQuerier, valAddr.String(), babyStakers); err != nil {
-			return nil, fmt.Errorf("failed to get delegations for validator %s: %w", valAddr.String(), err)
-		}
-	}
-
-	return babyStakers, nil
-}
-
-// getValidatorDelegations gets all delegations for a specific validator
-func getValidatorDelegations(ctx context.Context, stkQuerier stkkeeper.Querier, validatorAddr string, babyStakers map[string]math.Int) error {
-	var nextKey []byte
-
-	for {
-		req := &stktypes.QueryValidatorDelegationsRequest{
-			ValidatorAddr: validatorAddr,
-			Pagination: &query.PageRequest{
-				Key: nextKey,
-			},
+		if child.BtcUndelegation == nil || child.BtcUndelegation.DelegatorUnbondingInfo == nil {
+			return nil
 		}
 
-		res, err := stkQuerier.ValidatorDelegations(ctx, req)
+		// At this point the child is the GHSA-4rm2-cj74-f62h pattern:
+		// stake-expansion, no inclusion proof, but DelegatorUnbondingInfo set.
+		poisonedChildren++
+		childHashStr := child.MustGetStakingTxHash().String()
+
+		prevHash, err := child.StakeExpansionTxHash()
 		if err != nil {
-			return err
+			logger.Error("invalid stake-expansion previous tx hash on poisoned child — aborting upgrade",
+				"child_staking_tx_hash", childHashStr,
+				"error", err.Error())
+
+			return fmt.Errorf("invalid stake-expansion previous tx hash on child %s: %w",
+				childHashStr, err)
+		}
+		parentHashStr := prevHash.String()
+
+		parent, err := bk.GetBTCDelegation(ctx, parentHashStr)
+		if err != nil {
+			// Parent not in store — nothing to remediate. Could happen if the
+			// parent was already pruned by a future feature; safe to skip but
+			// unusual enough to warrant a Warn so operators notice.
+			skippedParentMissing++
+			logger.Warn("found poisoned child but parent delegation is not in store, skipping",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr,
+				"error", err.Error())
+
+			return nil //nolint:nilerr
+		}
+		if parent.BtcUndelegation != nil && parent.BtcUndelegation.DelegatorUnbondingInfo != nil {
+			// Already unbonded — either remediated previously or unbonded
+			// through a legitimate path that succeeded.
+			skippedParentAlreadyUnbond++
+			logger.Debug("found poisoned child but parent already unbonded, skipping",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr)
+
+			return nil
 		}
 
-		for _, delegation := range res.DelegationResponses {
-			delegatorAddr := delegation.Delegation.DelegatorAddress
-			amount := delegation.Balance.Amount
+		// Defense-in-depth: only remediate when the child's previous-stake
+		// covenant quorum was collected. Without `PreviousStkCovenantSigs`
+		// satisfying the parent's CovenantQuorum, the child cannot have
+		// produced a valid witness on TxIn[0] (which spends the parent's
+		// staking output via the parent's unbonding-path script), so the
+		// parent's UTXO was never actually spent on Bitcoin and the child's
+		// SpendStakeTx is not a true unbonding proof. Force-unbonding the
+		// parent in that case would remove voting power from a delegation
+		// whose UTXO is still alive on Bitcoin.
+		//
+		// The quorum is resolved from the PARENT's own ParamsVersion: the
+		// covenant committee that signs the parent's unbonding path is the
+		// committee that was active at parent registration (locked into the
+		// staking script), not whatever the chain-current params say.
+		parentParams := bk.GetParamsByVersion(ctx, parent.ParamsVersion)
+		if parentParams == nil {
+			skippedParentMissing++
+			logger.Warn("found poisoned child but parent ParamsVersion is unknown, skipping",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr,
+				"parent_params_version", parent.ParamsVersion)
 
-			if existing, found := babyStakers[delegatorAddr]; found {
-				babyStakers[delegatorAddr] = existing.Add(amount)
-			} else {
-				babyStakers[delegatorAddr] = amount
-			}
+			return nil
+		}
+		if child.StkExp == nil || !child.StkExp.HasCovenantQuorums(parentParams.CovenantQuorum) {
+			skippedNoPrevStkQuorum++
+			logger.Warn("found poisoned child but PreviousStkCovenantSigs do not meet parent quorum, skipping — parent UTXO cannot have been spent",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr,
+				"prev_stk_covenant_quorum", parentParams.CovenantQuorum)
+
+			return nil
 		}
 
-		if res.Pagination == nil || len(res.Pagination.NextKey) == 0 {
-			break
+		// Skip parents that are no longer ACTIVE for any reason (e.g.,
+		// their staking timelock expired and status became EXPIRED). The
+		// phantom-voting-power harm we are remediating only applies while
+		// the parent is contributing voting power; force-unbonding an
+		// already-zero-power parent would enqueue a redundant
+		// EventBTCDelegationStateUpdate{UNBONDED} on top of the prior
+		// expiry transition.
+		//
+		// IsBtcDelegationActive is the canonical "is this delegation
+		// ACTIVE?" check: it resolves the delegation's covenant quorum
+		// from its OWN ParamsVersion (not the chain-current params, which
+		// may have been bumped since the delegation was created), and for
+		// chained stake expansions also resolves the prev-stake's quorum
+		// from the prev-stake's own ParamsVersion. It returns a non-nil
+		// error for both real failures (delegation missing, params version
+		// missing) and the not-ACTIVE case; either way we skip the parent.
+		_, parentActive, _ := bk.IsBtcDelegationActive(ctx, parentHashStr)
+		if !parentActive {
+			skippedParentNotActive++
+			logger.Debug("found poisoned child but parent is not ACTIVE, skipping",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr)
+
+			return nil
 		}
-		nextKey = res.Pagination.NextKey
-	}
 
-	return nil
-}
+		if _, dup := seenParent[parentHashStr]; dup {
+			// Multiple registered stake expansions for the same ACTIVE parent
+			// are possible on babylon (registration only checks the parent is
+			// ACTIVE), but only one can ever confirm on Bitcoin. The recorded
+			// SpendStakeTx is best-effort — the first child encountered in
+			// iteration order. Voting power deduction works regardless of
+			// which one we pick.
+			skippedDuplicateParent++
+			logger.Warn("multiple poisoned children point to the same parent — only one stake expansion can spend the parent's UTXO on Bitcoin, skipping duplicate",
+				"child_staking_tx_hash", childHashStr,
+				"parent_staking_tx_hash", parentHashStr)
 
-// updateCostakerActiveBabyRewardsTracker creates or updates a costaker rewards tracker with ActiveBaby
-func updateCostakerActiveBabyRewardsTracker(
-	ctx context.Context,
-	coStkKeeper costkkeeper.Keeper,
-	endedPeriod uint64,
-	rwdTrackers collections.Map[[]byte, costktypes.CostakerRewardsTracker],
-	stakerAddr sdk.AccAddress,
-	params costktypes.Params,
-	babyTracked ActiveBabyTracked,
-) error {
-	addrKey := []byte(stakerAddr)
+			return nil
+		}
+		seenParent[parentHashStr] = struct{}{}
 
-	// Try to get existing tracker
-	rt, err := rwdTrackers.Get(ctx, addrKey)
-	if err != nil && !errors.Is(err, collections.ErrNotFound) {
-		return err
-	}
+		logger.Info("queued poisoned parent for remediation",
+			"parent_staking_tx_hash", parentHashStr,
+			"child_staking_tx_hash", childHashStr)
 
-	if errors.Is(err, collections.ErrNotFound) {
-		// This should not happen as we're updating existing trackers only
+		toRemediate = append(toRemediate, pair{
+			parent:     parent,
+			spendTxBz:  child.StakingTx,
+			childHash:  childHashStr,
+			parentHash: parentHashStr,
+		})
+
 		return nil
+	}); err != nil {
+		return nil, err
 	}
 
-	// Update existing tracker (set the ActiveBaby because it was zeroed out before)
-	// Update the StartPeriodCumulativeReward only if the ActiveBaby is changing
-	rt.ActiveBaby = rt.ActiveBaby.Add(babyTracked.CurrentActiveBaby)
-	diff := babyTracked.CurrentActiveBaby.Sub(babyTracked.PreviousActiveBaby)
-	if !diff.IsZero() { // if there is any diff, the period needs to be increased
-		rt.StartPeriodCumulativeReward = endedPeriod
-		if err := coStkKeeper.CalculateCostakerRewardsAndSendToGauge(ctx, stakerAddr, endedPeriod); err != nil {
-			return err
-		}
+	logger.Info("scan complete",
+		"delegations_scanned", scanned,
+		"poisoned_children_found", poisonedChildren,
+		"to_remediate", len(toRemediate),
+		"skipped_parent_missing", skippedParentMissing,
+		"skipped_parent_already_unbonded", skippedParentAlreadyUnbond,
+		"skipped_parent_not_active", skippedParentNotActive,
+		"skipped_duplicate_parent", skippedDuplicateParent,
+		"skipped_no_prev_stk_quorum", skippedNoPrevStkQuorum,
+	)
+
+	if len(toRemediate) == 0 {
+		logger.Info("no poisoned stake-expansion parents to remediate")
+
+		return nil, nil
 	}
 
-	// Update score
-	rt.UpdateScore(params.ScoreRatioBtcByBaby)
-
-	// Save tracker
-	if err := rwdTrackers.Set(ctx, addrKey, rt); err != nil {
-		return err
+	remediated := make([]string, 0, len(toRemediate))
+	for _, p := range toRemediate {
+		bk.BtcUndelegate(ctx, p.parent, &btcstktypes.DelegatorUnbondingInfo{
+			SpendStakeTx: p.spendTxBz,
+		})
+		ctx.EventManager().EmitEvent(sdk.NewEvent(
+			EventTypePhantomStakeExpansionRemediated,
+			sdk.NewAttribute(AttrParentStakingTxHash, p.parentHash),
+			sdk.NewAttribute(AttrChildStakingTxHash, p.childHash),
+		))
+		logger.Info("force-unbonded poisoned parent",
+			"parent_staking_tx_hash", p.parentHash,
+			"child_staking_tx_hash", p.childHash)
+		remediated = append(remediated, p.parentHash)
 	}
 
-	return nil
-}
+	logger.Info("remediation complete", "remediated_parents", len(remediated))
 
-// zeroOutCoStakerRwdsActiveBaby zeros out ActiveBaby in all costaker rewards trackers
-func zeroOutCoStakerRwdsActiveBaby(
-	ctx context.Context,
-	rwdTrackers collections.Map[[]byte, costktypes.CostakerRewardsTracker],
-) (map[string]ActiveBabyTracked, error) {
-	accsWithActiveBaby := make(map[string]ActiveBabyTracked)
-	iter, err := rwdTrackers.Iterate(ctx, nil)
-	if err != nil {
-		return accsWithActiveBaby, err
-	}
-	defer iter.Close()
-
-	for ; iter.Valid(); iter.Next() {
-		costakerAddr, err := iter.Key()
-		if err != nil {
-			return accsWithActiveBaby, err
-		}
-
-		tracker, err := iter.Value()
-		if err != nil {
-			return accsWithActiveBaby, err
-		}
-		if tracker.ActiveBaby.IsZero() {
-			continue
-		}
-
-		sdkAddr := sdk.AccAddress(costakerAddr)
-		accsWithActiveBaby[sdkAddr.String()] = ActiveBabyTracked{
-			PreviousActiveBaby: tracker.ActiveBaby,
-			CurrentActiveBaby:  math.ZeroInt(),
-		}
-
-		// Zero out ActiveBaby
-		tracker.ActiveBaby = math.ZeroInt()
-		if err := rwdTrackers.Set(ctx, costakerAddr, tracker); err != nil {
-			return accsWithActiveBaby, err
-		}
-	}
-
-	return accsWithActiveBaby, nil
-}
-
-func getTotalScore(
-	ctx context.Context,
-	rwdTrackers collections.Map[[]byte, costktypes.CostakerRewardsTracker],
-) (math.Int, error) {
-	totalScore := math.ZeroInt()
-	iter, err := rwdTrackers.Iterate(ctx, nil)
-	if err != nil {
-		return totalScore, err
-	}
-	defer iter.Close()
-
-	for ; iter.Valid(); iter.Next() {
-		tracker, err := iter.Value()
-		if err != nil {
-			return totalScore, err
-		}
-
-		totalScore = totalScore.Add(tracker.TotalScore)
-	}
-
-	return totalScore, nil
+	return remediated, nil
 }
